+++ packages/fxa-content-server/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.146.3
+
+### Bug fixes
+
+-   content/payment: update helmet (24fe65c44)
+
+### Other changes
+
+-   experiment: Enable sign-up codes 100% and sign-in codes 50% for sync users (375e1f5dd)
+
 ## 1.146.2
 
 ### Bug fixes
@@ -36,6 +46,12 @@
 
 -   docs: Add some missing fields in webchannel protocol docs (aa98940bd)
 -   settings: change "last sync" to "last seen" in devices view (04ae2e6f7)
+
+## 1.145.5
+
+### Bug fixes
+
+-   content/payment: update helmet (24fe65c44)
 
 ## 1.145.4