+++ packages/fxa-customs-server/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.146.3
+
+No changes.
+
 ## 1.146.2
 
 No changes.
@@ -12,6 +16,10 @@ No changes.
 
 - deps: fxa-customs-server npm audit changes (35e53017e)
 - deps: Update @google-cloud/pubsub for node 12 compat (913e67f6b)
+
+## 1.145.5
+
+No changes.
 
 ## 1.145.4