+++ packages/fxa-email-service/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.146.3
+
+No changes.
+
 ## 1.146.2
 
 No changes.
@@ -7,6 +11,10 @@ No changes.
 No changes.
 
 ## 1.146.0
+
+No changes.
+
+## 1.145.5
 
 No changes.